add a first pass at supporting ValueNet
@brunners does a nice job publishing ValueNet at https://github.com/brunnurs/valuenet.
It is fun to play with, since like IRNet it can deal with multiple tables,
and unlike IRNet it will take a shot at predicting values.  It is a little
tricky to run on fresh data that is not in the Spider dataset.  Out of curiosity,
I combined it with previous work on running IRNet on fresh data, and added a
docker image to support the same simple api as previous Sqlova and IRNet
experiments.

This was a quick test, and has various limits, including:
 * I made a quick and dirty replacement for use of the google entity
   recognition api, just to keep the test self-contained.
 * The sqlite-handling code seems to have a problem with columns with
   spaces in their names at the moment, oops.